Title: 낮은 지점들, Time: 78ms, Memory: 2MB, Status: Passed - CodeTree

+++ 240715/낮은 지점들/lowest-points.cpp
@@ -18,11 +18,11 @@ int main() {
         }
     }
 
-    int sum = 0;
+    long long sum = 0;
     for(auto i: m){
         sum += i.second;
     }
     
-    printf("%d", sum);
+    printf("%lld", sum);
     return 0;
 }
